Move Sender from workers to helpers
