+++ bee-protocol/src/worker/mod.rs
@@ -17,7 +17,6 @@ mod peer;
 mod propagator;
 mod requester;
 mod responder;
-mod sender;
 mod status;
 mod tps;
 mod transaction;
@@ -36,7 +35,6 @@ pub(crate) use responder::{
     MilestoneResponderWorker, MilestoneResponderWorkerEvent, TransactionResponderWorker,
     TransactionResponderWorkerEvent,
 };
-pub(crate) use sender::Sender;
 pub(crate) use status::StatusWorker;
 pub(crate) use tps::TpsWorker;
 pub(crate) use transaction::{HasherWorker, HasherWorkerEvent, ProcessorWorker};
